Reformatted events array
I added <br> instead of newline to the events array strings. (Or maybe I should've kept the \n and then parsed it on the other end. Oops.)

I also split up the lines so they were easier to read in the code.

I've also added an array of connections (in numeric form) to each event. It was just too difficult to keep track of without putting the info there.