Added all event connections
I listed all the connections between events as an array in each call to the constructor. I figured that was the easiest way to do it. It may give the event some information it doesn't need, but I wanted all the info in one place. (It makes the code more self-documenting.)